+++ functions.py
@@ -9,3 +9,12 @@ def minus(a, b):
 def multiply(a, b):
     """Multiply two numbers. """
     return a * b
+
+def exp(a, b):
+    """Exponential a**b."""
+    return a ** b
+
+def divide(a, b):
+    """Divide b from a"""
+    return a / b
+